Using size_t for vector dimension variables
Preventing the (very unlikely) overflow case

+++ src/adcirc/geometry/boundary.cpp
@@ -36,7 +36,7 @@ Boundary::Boundary() { this->setBoundary(-1, 0); }
  * @param boundaryCode ADCIRC model boundary code
  * @param boundaryLength number of nodes along this boundary
  */
-Boundary::Boundary(int boundaryCode, int boundaryLength) {
+Boundary::Boundary(int boundaryCode, size_t boundaryLength) {
   this->setBoundaryCode(boundaryCode);
   this->setBoundaryLength(boundaryLength);
 }
@@ -47,7 +47,7 @@ Boundary::Boundary(int boundaryCode, int boundaryLength) {
  * @param boundaryCode ADCIRC model boundary code
  * @param boundaryLength number of nodes along this boundary
  */
-void Boundary::setBoundary(int boundaryCode, int boundaryLength) {
+void Boundary::setBoundary(int boundaryCode, size_t boundaryLength) {
   this->setBoundaryCode(boundaryCode);
   this->setBoundaryLength(boundaryLength);
 }
@@ -57,7 +57,7 @@ void Boundary::setBoundary(int boundaryCode, int boundaryLength) {
  * @brief Returns the length of the boundary
  * @return number of nodes in boundary
  */
-int Boundary::boundaryLength() const { return this->m_boundaryLength; }
+size_t Boundary::boundaryLength() const { return this->m_boundaryLength; }
 
 /**
  * @name Boundary::size
@@ -65,7 +65,7 @@ int Boundary::boundaryLength() const { return this->m_boundaryLength; }
  * Boundary::bonudaryLength
  * @return number of nodes in boundary
  */
-int Boundary::size() const { return this->boundaryLength(); }
+size_t Boundary::size() const { return this->boundaryLength(); }
 
 /**
  * @name Boundary::length
@@ -73,14 +73,14 @@ int Boundary::size() const { return this->boundaryLength(); }
  * Boundary::bonudaryLength
  * @return number of nodes in boundary
  */
-int Boundary::length() const { return this->boundaryLength(); }
+size_t Boundary::length() const { return this->boundaryLength(); }
 
 /**
  * @name Boundary::setBoundaryLength
  * @brief Allocates the arrays used by the bounary
  * @param boundaryLength number of nodes along the boundary
  */
-void Boundary::setBoundaryLength(int boundaryLength) {
+void Boundary::setBoundaryLength(size_t boundaryLength) {
   if (this->boundaryLength() != boundaryLength) {
     this->m_boundaryLength = boundaryLength;
     this->m_node1.resize(this->boundaryLength());
@@ -129,7 +129,7 @@ void Boundary::setBoundaryCode(int boundaryCode) {
  * @param index position along the boundary
  * @return crest elevation if applicable, otherwise -9999.0.
  */
-double Boundary::crestElevation(int index) const {
+double Boundary::crestElevation(size_t index) const {
   if (this->boundaryCode() == 3 || this->boundaryCode() == 13 ||
       this->boundaryCode() == 23 || this->boundaryCode() == 4 ||
       this->boundaryCode() == 24 || this->boundaryCode() == 5 ||
@@ -146,7 +146,7 @@ double Boundary::crestElevation(int index) const {
  * @param index position along the boundary
  * @param crestElevation height above the datum for the weir crest
  */
-void Boundary::setCrestElevation(int index, double crestElevation) {
+void Boundary::setCrestElevation(size_t index, double crestElevation) {
   if (this->boundaryCode() == 3 || this->boundaryCode() == 13 ||
       this->boundaryCode() == 23 || this->boundaryCode() == 4 ||
       this->boundaryCode() == 24 || this->boundaryCode() == 5 ||
@@ -162,7 +162,7 @@ void Boundary::setCrestElevation(int index, double crestElevation) {
  * @param index position along the boundary
  * @return subcritical weir coefficient
  */
-double Boundary::subcriticalWeirCoefficient(int index) const {
+double Boundary::subcriticalWeirCoefficient(size_t index) const {
   if (this->boundaryCode() == 4 || this->boundaryCode() == 24 ||
       this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
@@ -176,8 +176,7 @@ double Boundary::subcriticalWeirCoefficient(int index) const {
  * @param index position along the boundary
  * @param subcriticalWeirCoefficient coefficient of subcritical flow
  */
-void Boundary::setSubcriticalWeirCoefficient(
-    int index, double subcriticalWeirCoefficient) {
+void Boundary::setSubcriticalWeirCoefficient(size_t index, double subcriticalWeirCoefficient) {
   if (this->boundaryCode() == 4 || this->boundaryCode() == 24 ||
       this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
@@ -191,7 +190,7 @@ void Boundary::setSubcriticalWeirCoefficient(
  * @param index position along the boundary
  * @return coefficient of supercritical flow
  */
-double Boundary::supercriticalWeirCoefficient(int index) const {
+double Boundary::supercriticalWeirCoefficient(size_t index) const {
   if (this->boundaryCode() == 3 || this->boundaryCode() == 13 ||
       this->boundaryCode() == 23 || this->boundaryCode() == 4 ||
       this->boundaryCode() == 24 || this->boundaryCode() == 5 ||
@@ -208,7 +207,7 @@ double Boundary::supercriticalWeirCoefficient(int index) const {
  * boundarytypes 3, 13, 23, 4, 24, 5, and 25
  */
 void Boundary::setSupercriticalWeirCoefficient(
-    int index, double supercriticalWeirCoefficient) {
+    size_t index, double supercriticalWeirCoefficient) {
   if (this->boundaryCode() == 3 || this->boundaryCode() == 13 ||
       this->boundaryCode() == 23 || this->boundaryCode() == 4 ||
       this->boundaryCode() == 24 || this->boundaryCode() == 5 ||
@@ -225,7 +224,7 @@ void Boundary::setSupercriticalWeirCoefficient(
  * @param index position along the boundary
  * @return height of pipe center above datum
  */
-double Boundary::pipeHeight(int index) const {
+double Boundary::pipeHeight(size_t index) const {
   if (this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
       return this->m_pipeHeight[index];
@@ -239,7 +238,7 @@ double Boundary::pipeHeight(int index) const {
  * @param index position along the boundary
  * @param pipeHeight elevation of the pipe center above datum
  */
-void Boundary::setPipeHeight(int index, double pipeHeight) {
+void Boundary::setPipeHeight(size_t index, double pipeHeight) {
   if (this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
       this->m_pipeHeight[index] = pipeHeight;
@@ -251,7 +250,7 @@ void Boundary::setPipeHeight(int index, double pipeHeight) {
  * @param index position along the boundary
  * @return diameter of the pipe
  */
-double Boundary::pipeDiameter(int index) const {
+double Boundary::pipeDiameter(size_t index) const {
   if (this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
       return this->m_pipeDiameter[index];
@@ -264,7 +263,7 @@ double Boundary::pipeDiameter(int index) const {
  * @param index postion along the boundary
  * @param pipeDiameter diameter of the pipe
  */
-void Boundary::setPipeDiameter(int index, double pipeDiameter) {
+void Boundary::setPipeDiameter(size_t index, double pipeDiameter) {
   if (this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
       this->m_pipeDiameter[index] = pipeDiameter;
@@ -276,7 +275,7 @@ void Boundary::setPipeDiameter(int index, double pipeDiameter) {
  * @param index position along the boundary
  * @return pipe coefficient
  */
-double Boundary::pipeCoefficient(int index) const {
+double Boundary::pipeCoefficient(size_t index) const {
   if (this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
       return this->m_pipeCoefficient[index];
@@ -289,7 +288,7 @@ double Boundary::pipeCoefficient(int index) const {
  * @param index position along boundary
  * @param pipeCoefficient pipe coefficient
  */
-void Boundary::setPipeCoefficient(int index, double pipeCoefficient) {
+void Boundary::setPipeCoefficient(size_t index, double pipeCoefficient) {
   if (this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
       this->m_pipeCoefficient[index] = pipeCoefficient;
@@ -301,7 +300,7 @@ void Boundary::setPipeCoefficient(int index, double pipeCoefficient) {
  * @param index position along the boundary
  * @return pointer to an Node object
  */
-Node *Boundary::node1(int index) const {
+Node *Boundary::node1(size_t index) const {
   if (index < this->boundaryLength() || index >= 0) return this->m_node1[index];
   return nullptr;
 }
@@ -312,7 +311,7 @@ Node *Boundary::node1(int index) const {
  * @param index position along the boundary
  * @param node1 Pointer to an Node object
  */
-void Boundary::setNode1(int index, Node *node1) {
+void Boundary::setNode1(size_t index, Node *node1) {
   if (index < this->boundaryLength() || index >= 0)
     this->m_node1[index] = node1;
 }
@@ -324,7 +323,7 @@ void Boundary::setNode1(int index, Node *node1) {
  * @param index position along the boundary
  * @return Node pointer
  */
-Node *Boundary::node2(int index) const {
+Node *Boundary::node2(size_t index) const {
   if (this->boundaryCode() == 4 || this->boundaryCode() == 24 ||
       this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
@@ -338,7 +337,7 @@ Node *Boundary::node2(int index) const {
  * @param index position along the boundary
  * @param node2 pointer to an Node object
  */
-void Boundary::setNode2(int index, Node *node2) {
+void Boundary::setNode2(size_t index, Node *node2) {
   if (this->boundaryCode() == 4 || this->boundaryCode() == 24 ||
       this->boundaryCode() == 5 || this->boundaryCode() == 25)
     if (index < this->boundaryLength() || index >= 0)
